+++ src/main/java/com/dimasukimas/tennisscoreboard/service/scoring/ScoringStrategyFactory.java
@@ -1,13 +1,13 @@
 package com.dimasukimas.tennisscoreboard.service.scoring;
 
-import com.dimasukimas.tennisscoreboard.enums.ScoringStrategyType;
-import com.dimasukimas.tennisscoreboard.model.match.OngoingMatch;
+import com.dimasukimas.tennisscoreboard.enumeration.ScoringStrategyType;
+import com.dimasukimas.tennisscoreboard.model.common.OngoingMatch;
 
 public class ScoringStrategyFactory {
 
-    private static final ScoringStrategy <OngoingMatch, Integer> DEFAULT_STRATEGY = new BestOfThreeScoringStrategy();
+    private static final ScoringStrategy <OngoingMatch, Long> DEFAULT_STRATEGY = new BestOfThreeScoringStrategy();
 
-    public static ScoringStrategy<OngoingMatch, Integer> getStrategy(ScoringStrategyType type) {
+    public static ScoringStrategy<OngoingMatch, Long> getStrategy(ScoringStrategyType type) {
         return DEFAULT_STRATEGY;
     }
 }
